Update README.md
Remember that the human lifespan extended after the problem of heating and access to clean water was resolved. Do you want a long life? Solve the problem of clean water and heating. This topic also contains the solution of this work.
We can't reach clean water in near future free. But there is a way for it not so expensive and by solar energy.
If we want to be prepared to live in future, we will need some energy producing systems and clean water for usage and irrigation our crops.
Besides those, these assets should be autonomous and work by natural resources. This subject provides all using by solar power.
You can get clean water from swamps, polluted rivers, lakes or seas.